Allow creation of DomainEventProcessor from function handles
This may seem like unnecessary indirection due to the way we unwrap
these DomainEventProcessors inside of EventProcessor, however this makes
us consistent with the way we do something similar over on the write
(Aggregate and Route) side, i.e. DomainEventProcessors are like
Aggregates, and EventProcessors are like Routes in that they wrap and
perform a routing function including the reified type magic it needs to
do in order to do its job.