Use blocking lock function in sink function to avoid sequence gap issue
As detailed in DemonstrateEventSequenceIdGaps.kt, there is a concurrency
race condition wherein it's possible for the event-processors to skip an
event. Using a blocking lock on the write side works around that
problem. We will be 1/2 to 2/3 slower to sink events, but we're a lot
safer so we can probably wear that issue.

PR where Odin fixes this for the in-house Ruby event framework using the
same blocking technique. Our fix is copied from his:
https://github.com/cultureamp/event_framework/pull/20

Threads where we talked about this issue:
https://cultureamp.slack.com/archives/CGANLQ5L5/p1583807707015300?thread_ts=1583802150.013000&cid=CGANLQ5L5
https://cultureamp.slack.com/archives/C017MPV2Q1X/p1602820361023400?thread_ts=1602737223.007500&cid=C017MPV2Q1X

Event-sourcery dealing with demonstrating problem via a script that
we've shamelessly copied:
https://github.com/envato/event_sourcery-postgres/blob/8bd15a28c277ee590fc823ffdd59931ef4614d33/script/demonstrate_event_sequence_id_gaps.rb