Allow table lock timeout to be configured
Allows a scenario where you want to allow non-blocking clients to wait and come back and retry manually. This can be useful if the events table has high contention, and you want to be able to privilege some callers over others.

Still defaults to 10sec by default, which can be overridden without having to override the whole locking function.

Also opened up class constructor because so that it's easier to experiment with changes like this without having to release new Kestrel versions.